requirements.txt
django~=4.2.5
djangorestframework~=3.14.0
django-jazzmin~=2.6.0
environs~=9.5.0
requests~=2.31.0
aiogram~=2.9